refactor: 위젯 편집 페이지로 이동
