+++ app/(home)/_components/WidgetItem.tsx
