+++ components/ui/Button.tsx
@@ -2,7 +2,7 @@ import { cn } from '@/utils/twMerge';
 import { VariantProps, cva } from 'class-variance-authority';
 import { ButtonHTMLAttributes, forwardRef } from 'react';
 
-export const buttonVariants = cva('', {
+export const buttonVariants = cva('flex items-center justify-center', {
   variants: {
     size: {
       xs: 'h-[3.2rem] rounded-xs px-16 text-12',
